fragments2pdf/requirements.txt
pyairtable==2.2.2
python-dotenv==1.0.1
jinja2==3.1.3
xhtml2pdf==0.2.13
pdfminer.six==20231228
reportlab==4.0.9
markdown==3.5.2
tqdm==4.66.4